refactor(snapshot): one file per BlockTools implementation
block_tools.rs held the seam and its only implementation. A second one
is coming for stock-distro userlands, and what differs between them is
the whole tool surface — busybox drives one multi-call binary with
applet names, util-linux two separate programs with long options — so
they get a file each rather than a configurable path.

mod.rs keeps the trait and what both implementations share: the
`/dev/loopN` parser, the sysfs attach verification (now taking the tools
it may have to detach through), the device-node helpers, and the fake
binary the tests drive.
